requirements.txt
PyMuPDF
pandas
torch
argparse
transformers
tqdm
argparse
scikit-learn
utils
matplotlib
numpy
seqeval
grpcio
grpcio-tools
minio
tensorflow
nltk